+++ Source/Types/DryRunResult.swift
@@ -1,0 +1,13 @@
+//
+//  DryRunResult.swift
+//  CKB
+//
+//  Created by James Chen on 2019/05/13.
+//  Copyright © 2019 Nervos Foundation. All rights reserved.
+//
+
+import Foundation
+
+public struct DryRunResult: Codable {
+    public let cycles: Cycle
+}