+++ Source/Types/Header.swift
@@ -14,16 +14,17 @@ public struct Seal: Codable {
 }
 
 public struct Header: Codable {
-    public let version: UInt32
+    public let version: Version
     public let parentHash: H256
-    public let timestamp: String
+    public let timestamp: Timestamp
     public let number: BlockNumber
+    public let epoch: EpochNumber
     public let transactionsRoot: H256
-    public let proposalsRoot: H256
+    public let proposalsHash: H256
     public let witnessesRoot: H256
-    public let difficulty: UInt256
+    public let difficulty: HexNumber
     public let unclesHash: H256
-    public let unclesCount: UInt32
+    public let unclesCount: Number
     public let hash: H256
 
     public let seal: Seal
@@ -33,8 +34,9 @@ public struct Header: Codable {
         case parentHash = "parent_hash"
         case timestamp
         case number
+        case epoch
         case transactionsRoot = "transactions_root"
-        case proposalsRoot = "proposals_root"
+        case proposalsHash = "proposals_hash"
         case witnessesRoot = "witnesses_root"
         case difficulty
         case unclesHash = "uncles_hash"