+++ Source/Types/OutPoint.swift
@@ -9,15 +9,41 @@
 import Foundation
 
 public struct OutPoint: Codable, Param {
+    public let blockHash: H256?
+    public let cell: CellOutPoint?
+
+    enum CodingKeys: String, CodingKey {
+        case blockHash = "block_hash"
+        case cell
+    }
+
+    public init(blockHash: H256? = nil, cell: CellOutPoint) {
+        self.blockHash = blockHash
+        self.cell = cell
+    }
+
+    public var param: [String: Any] {
+        var result = [String: Any]()
+        if let blockHash = blockHash {
+            result["block_hash"] = blockHash
+        }
+        if let cell = cell {
+            result["cell"] = cell.param
+        }
+        return result
+    }
+}
+
+public struct CellOutPoint: Codable, Param {
     public let txHash: H256
-    public let index: UInt32
+    public let index: Number
 
     enum CodingKeys: String, CodingKey {
         case txHash = "tx_hash"
         case index
     }
 
-    public init(txHash: H256, index: UInt32) {
+    public init(txHash: H256, index: Number) {
         self.txHash = txHash
         self.index = index
     }