+++ Tests/API/APIMockingTests.swift
@@ -28,6 +28,16 @@ class APIMockingTests: XCTestCase {
         XCTAssertEqual("0", result.header.number)
     }
 
+    func testGetCurrentEpoch() throws {
+        let result = try getClient(json: "epoch").getCurrentEpoch()
+        XCTAssertTrue(UInt64(result.difficulty.dropFirst(2), radix: 16)! >= 0)
+    }
+
+    func testGetEpochByNumber() throws {
+        let result = try getClient(json: "epoch").getEpochByNumber(number: "0")
+        XCTAssertNotNil(result)
+    }
+
     func testGetTransaction() throws {
         let result = try getClient(json: "transaction").getTransaction(hash: "0x2505abd12b6353da33152014cabdf68566fea3976986b5da2fd5980940191ef5")
         XCTAssertNotNil(result)
@@ -45,7 +55,10 @@ class APIMockingTests: XCTestCase {
     }
 
     func testGetLiveCell() throws {
-        let outPoint = OutPoint(txHash: "0xda27a795e14067e18a0cfc0951571aaca47bd9851adc9bf5baa22cf27c8bcde8", index: 0)
+        let outPoint = OutPoint(
+            blockHash: "0x4c2f8ba5f5a0104eaf84fcbb16af4b0e7ca2f2fdb076e748d54ef876d085d49e",
+            cell: CellOutPoint(txHash: "0xda27a795e14067e18a0cfc0951571aaca47bd9851adc9bf5baa22cf27c8bcde8", index: "0")
+        )
         let result = try getClient(json: "liveCellEmpty").getLiveCell(outPoint: outPoint)
         XCTAssertNotNil(result)
     }
@@ -61,15 +74,45 @@ class APIMockingTests: XCTestCase {
         XCTAssertNil(result)
     }
 
+    func testTxPoolInfo() throws {
+        let result = try getClient(json: "txPoolInfo").txPoolInfo()
+        XCTAssertNotNil(result)
+        XCTAssert(UInt32(result.pending)! >= 0)
+    }
+
+    func testGetBlockchainInfo() throws {
+        let result = try getClient(json: "blockchainInfo").getBlockchainInfo()
+        XCTAssertNotNil(result)
+        XCTAssertFalse(result.chain.isEmpty)
+    }
+
+    func testGetPeersState() throws {
+        let result = try getClient(json: "peersState").getPeersState()
+        XCTAssertNotNil(result)
+        XCTAssert(result.count >= 0)
+    }
+
     func testLocalNodeInfo() throws {
         let result = try getClient(json: "localNodeInfo").localNodeInfo()
         XCTAssertFalse(result.addresses.isEmpty)
         XCTAssertFalse(result.nodeId.isEmpty)
     }
 
-    func testTraceTransactionEmpty() throws {
-        let tx = Transaction(deps: [], inputs: [], outputs: [], witnesses: [])
-        let result = try? getClient(json: "traceTransactionEmpty").traceTransaction(transaction: tx)
+    func testGetPeers() throws {
+        let result = try getClient(json: "peers").getPeers()
+        XCTAssertNotNil(result)
+        XCTAssertFalse(result.first!.addresses.isEmpty)
+    }
+
+    func testComputeTransactionHash() throws {
+        let tx = Transaction()
+        let result = try? getClient(json: "computeTransactionHash").computeTransactionHash(transaction: tx)
+        XCTAssertNotNil(result)
+    }
+    
+    func testDryRunTransaction() throws {
+        let tx = Transaction()
+        let result = try? getClient(json: "dryRunTransaction").dryRunTransaction(transaction: tx)
         XCTAssertNotNil(result)
     }
 }
@@ -78,8 +121,7 @@ extension APIMockingClient {
     convenience init(_ jsonPath: String) {
         self.init(url: APIClient.defaultLocalURL)
 
-        let path = Bundle(for: type(of: self)).path(forResource: jsonPath, ofType: "json")!
-        mockingData = try? Data(contentsOf: URL(fileURLWithPath: path), options: .mappedIfSafe)
+        mockingData = TestHelper.load(json: jsonPath)
     }
 }