+++ Tests/Types/ScriptTests.swift
@@ -10,17 +10,17 @@ import XCTest
 @testable import CKB
 
 class ScriptTests: XCTestCase {
-    func testEmptyScriptTypeHash() {
+    func testEmptyScriptHash() {
         let script = Script()
-        XCTAssertEqual("0x266cec97cbede2cfbce73666f08deed9560bdf7841a7a5a51b3a3f09da249e21", script.typeHash)
+        XCTAssertEqual("0x266cec97cbede2cfbce73666f08deed9560bdf7841a7a5a51b3a3f09da249e21", script.hash)
     }
 
-    func testScriptTypeHash() {
+    func testScriptHash() {
         let script = Script(
             args: ["0x01"],
             codeHash: H256.zeroHash
         )
-        XCTAssertEqual("0xdade0e507e27e2a5995cf39c8cf454b6e70fa80d03c1187db7a4cb2c9eab79da", script.typeHash)
+        XCTAssertEqual("0xdade0e507e27e2a5995cf39c8cf454b6e70fa80d03c1187db7a4cb2c9eab79da", script.hash)
     }
 
     func testCodeHashHasPrefix() {